requirements.txt
django>=2.2,<2.3
djangorestframework>=3.10,<3.11
pycryptodome
sentry-sdk==0.16.2
raven
